refactor: extract search to its own module
